refactor(Either): split out operations and dsls
