+++ packages/core/src/Either/dsls/matchers.ts
@@ -1,0 +1,10 @@
+// ets_tracing: off
+
+import * as DSL from "../../Prelude/DSL"
+import { Covariant } from "../instances"
+
+/**
+ * Matchers
+ */
+export const { match, matchIn, matchMorph, matchTag, matchTagIn } =
+  DSL.matchers(Covariant)
