+++ packages/core/src/Either/dsls/sequence.ts
@@ -1,0 +1,6 @@
+// ets_tracing: off
+
+import { sequenceF } from "../../Prelude/ForEach"
+import { ForEach } from "../instances"
+
+export const sequence = sequenceF(ForEach)
