+++ packages/core/src/Either/dsls/struct.ts
@@ -1,0 +1,6 @@
+// ets_tracing: off
+
+import * as DSL from "../../Prelude/DSL"
+import { Applicative } from "../instances"
+
+export const struct = DSL.structF(Applicative)
