+++ packages/core/src/Either/dsls/tuple.ts
@@ -1,0 +1,6 @@
+// ets_tracing: off
+
+import * as DSL from "../../Prelude/DSL"
+import { Applicative } from "../instances"
+
+export const tuple = DSL.tupleF(Applicative)
